+++ CHANGES.rst
@@ -8,6 +8,9 @@ Changelog
 
 -   Retry email bootstep when IMAP connection is reset.
 
+-   Change Bilby accounting group user to cody.messick as temporary workaround
+    while KAGRA accounts are in flux.
+
 2.3.2 "Champ" (2024-03-22)
 --------------------------