+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+2.3.3 (unreleased)
+------------------
+
+-   Change Bilby accounting group user to cody.messick as temporary workaround
+    while KAGRA accounts are in flux.
+
 2.3.2 "Champ" (2024-03-22)
 --------------------------